img_classification/requirements.txt
numpy==1.23.5
tensorflow==2.14.0
pillow==10.3.0

